+++ src/main/java/com/contract/management/system/dao/contract/ContractProductDao.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ContractProductDao
 {
     ContractProductEntity save(ContractProductEntity entity);
-    void deleteById(int id);
-    ContractProductEntity findById(int id);
+    void deleteById(long id);
+    ContractProductEntity findById(long id);
     List<ContractProductEntity> findAll();
 }
