+++ src/main/java/com/contract/management/system/model/collateral/dto/CollateralDto.java
@@ -13,12 +13,12 @@ import java.io.Serializable;
 @AllArgsConstructor
 public class CollateralDto implements Serializable
 {
-    private int collateralId;
+    private long collateralId;
     private String code;
     private String name;
     private int insurableMoney;
     private int standardMoney;
-    private int productId;
+    private long productId;
     private String productCode;
     private String productName;
 }
