+++ src/main/java/com/contract/management/system/repository/ContractCollateralRepository.java
@@ -3,5 +3,5 @@ package com.contract.management.system.repository;
 import com.contract.management.system.model.contract.entity.ContractCollateralEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContractCollateralRepository extends JpaRepository<ContractCollateralEntity, Integer>
+public interface ContractCollateralRepository extends JpaRepository<ContractCollateralEntity, Long>
 {}
