+++ src/main/java/com/contract/management/system/repository/ContractRepository.java
@@ -3,7 +3,7 @@ package com.contract.management.system.repository;
 import com.contract.management.system.model.contract.entity.ContractEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContractRepository extends JpaRepository<ContractEntity, Integer>
+public interface ContractRepository extends JpaRepository<ContractEntity, Long>
 {
 
 }
