+++ src/main/java/com/contract/management/system/repository/ProductRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends JpaRepository<ProductEntity, Integer>
+public interface ProductRepository extends JpaRepository<ProductEntity, Long>
 {
     List<ProductEntity> findByNameLike(String name);
 }
